Setting default language using env vars
Setting the env var "W4_LANG" changes the default value for "new" and "png2src".
For this, a new option was added: "--lang".
This option can only have these options:

- as
- assemblyscript
- c
- rs
- rust
- go

In case no value was provided by either env or cli,
it falls back to "as".
It can still be overriden by using the options --as etc.