Added "w4 init" - Fixes #50
In essence a "w4 init" is an alias to "w4 new ."

To achieve this and reduce code duplications, I've created a new
function called "blankProject" inside cli.js.
It takes a program.command and attaches the options of the "new" command
This function then is called for "new" and "init".

To make it a bit simpler still, I've also added a "real" alias to
"new" called "create".

Further "changes"to the source code are due to auto-formating it.