Adding custom ports to run and watch
The new `port` option allows to specify one or more ranges for the port.
The default is a range between 4000 and 6000
A range can be specified like `4000-6000`.
A fixed port can also be specified: --port 4444
It's also possible to provide multiple ranges:
`--port 4000-6000,8000-7000,8080`
If a port is taken, it attempts to do this again.
After 10 attempts, an exception is thrown.

This Solves #198