Increase wasm3 stack size from 4 KB to 64 KB.
This is an arbitrary limit corresponding to the implementation details
of the wasm3 interpreter. It's unrelated to the resource constraints of
the wasm4 VM. Making this too small will ultimately result in `[trap]
stack overflow` at the command line.

Using 64 KB, since this is the default of wasm3 standalone binary on
desktop platforms (from wasm3/platforms/app/main.c).

Fixes https://github.com/aduros/wasm4/issues/235