Remove text null-terminator discrepancy
On web, calls to `textUtf8` and `textUtf16` will stop at a
null-terminator even when given a length. Native does not do this,
but instead will happily try to print the null character despite there
being no font glyph for it.

This commit changes native to also stop at null-terminators to remove
the discrepancy between the runtimes.

We could also instead change web to print out the null characters, but that
doesn't serve any purpose and is less likely to be what a game creator wants.