APU: Make all four ADSR sections tick based
Each tone call is split into sections, each of which has a start volume and
frequency, and an end volume and frequency. Each section is played
continuously until either its end *tick* is reached, ending the section and
starting the next, or its end *time* is reached, which holds the tone there
and stretches it out until its end tick is reached.