Remove node types from web runtime
The web runtime doesn't run on NodeJS, but was using node types. This
was causing issues because some node types override the DOM types,
e.g. `setTimeout()`.

Adding `"types": []` to tsconfig.json filters out the types in
`node_modules/@types` and prevents them from being used, including
`node_modules/@types/node`, which is installed because it is a
dependency of our dependencies.