Rework update() pacing to be regular and work in background
This commit reworks the frame timing of the runtime ensuring a regular update
every frame on 60 fps monitors, while still supporting other framerates both
higher and lower. It achieves this by measuring the framerate continuously, and
updating in a requestAnimationFrame() callback when the framerate is close to
60, but switching to a setTimer() timing scheme otherwise.

This keeps animation smooth on a 60 fps screen, but keeps updates both
full-speed and at regular 16ms intervals on a 30 fps screen, important for
making tick-based audio timings accurate. This also has the effect of allowing
you to run the runtime in a minimised or background browser window. This is
really improves the audio experience, allows you to run games or demos in the
background, and also makes it much easier to test netplay without having to
keep many windows on screen at once.