Rework update() pacing to be regular and work in background
This commit reworks the frame timing of the runtime ensuring a regular
update on every frame on 60 fps monitors, while still supporting other
framerates, both higher and lower. It achieves this by measuring the
framerate continuously, and performing updates in a
requestAnimationFrame() callback when the framerate is close to 60, but
switching to a setTimeout() timing scheme otherwise.

This keeps animation smooth on a 60 fps screen, but keeps updates both
full-speed and at a regular 60Hz on e.g. a 30 fps screen, important for
keeping tick-based audio smooth. This also has the effect of allowing
you to run the runtime in a minimised or background browser window. This
really improves the audio experience, allowing you to run games or music
carts in the background. This also improves the netplay experience,
as the game won't lag or halt any time a player clicks out of the
browser window or changes tab.