make diskw skip localStorage.setItem for non-host players
recently i got surprised to see diskw during a netplay performs
localStorage.setItem even on non-host players. (and it overwrote
the save data for the cart i happened to have on the browser.)
as the disk buffer is a part of the shared state provided by the
netplay host, it's more natural to make diskw perform
localStorage.setItem only on the host.

cf. https://github.com/aduros/wasm4/issues/837